requirements.txt
Django==5.2.8
django-decouple==2.1
PyMySQL==1.1.1
cryptography==41.0.7
djangorestframework==3.15.2
django-cors-headers==4.6.0
Pillow==10.2.0
qrcode==7.4.2